Add files via upload
This repo contains my analysis on global heath statistics dataset, the dashboard file (pbix) link is in the "project report file", I put it there because the size upload data limit(file) is below the dashboard size. The user can download and view on Power BI desktop app. However, for a static view of the dashboard the user can view it through the PDF file attached in the repository